Factor out common parts of aggregation
