+++ lib/blocktrain.rb
@@ -8,6 +8,7 @@ require 'singleton'
 require 'blocktrain/version'
 require 'blocktrain/client'
 require 'blocktrain/lookups'
+require 'blocktrain/aggregation'
 require 'blocktrain/train_weight_aggregation'
 require 'blocktrain/car_weight_aggregation'
